+++ lib/channels.dart
@@ -4,6 +4,7 @@ export 'src/piesocket.dart';
 export 'src/channel.dart';
 export 'src/connection.dart';
 export 'src/auth_resolver.dart';
+export 'src/pie_rtc.dart';
 export 'src/misc/piesocket_event.dart';
 export 'src/misc/piesocket_exception.dart';
 export 'src/misc/piesocket_options.dart';